pyproject.toml
[project]
name = "realtime-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "llvmlite==0.40.1",
    "numba==0.57.1",
    "numpy>=1.21,<1.25",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.0.1",
    "resampy>=0.4.3",
    "soundfile>=0.12.1",
    "websockets>=13.1",
]
